# Scoreboard-backend/config/swagger.json
{
	"components": {
...
		"schemas": {
			"UserRole": {
				"enum": [
					"Admin",
					"ScoreEditor"
				],
				"type": "string"
			},
...
			"DeleteUserRequest": {
				"properties": {
					"username": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccessTokenTypes": {
				"enum": [
					"Basic",
					"Bearer"
				],
				"type": "string"
			},
			"CreateUserWithUsernameAndPasswordRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChessScoreUpdateRequest": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CricketScoreUpdateRequest": {
				"properties": {
					"teamA_runs": {
						"type": "number",
						"format": "double"
					},
					"teamA_overs": {
						"type": "number",
						"format": "double"
					},
					"teamA_wickets": {
						"type": "number",
						"format": "double"
					},
					"teamB_runs": {
						"type": "number",
						"format": "double"
					},
					"teamB_overs": {
						"type": "number",
						"format": "double"
					},
					"teamB_wickets": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_runs",
					"teamA_overs",
					"teamA_wickets",
					"teamB_runs",
					"teamB_overs",
					"teamB_wickets"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Winner": {
				"properties": {
					"team": {
						"type": "string"
					},
					"participants": {
						"items": {
							"$ref": "#/components/schemas/Participant"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"FootballEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/FootballScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChessEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/ChessScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CricketEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/CricketScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					},
					"eventLink": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams",
					"eventLink"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SquashMenEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/SquashMenScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SquashWomenEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/SquashWomenScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TennisMenEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/TennisMenScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					},
					"matchType": {
						"$ref": "#/components/schemas/MatchTypes"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams",
					"matchType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TennisWomenEvent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"subtitle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/EventCatagories"
					},
					"isStarted": {
						"type": "boolean"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"$ref": "#/components/schemas/TennisWomenScore"
					},
					"isCompleted": {
						"type": "boolean"
					},
					"teams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"participants": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Participant"
							},
							"type": "array"
						},
						"type": "array"
					},
					"votes": {
						"properties": {
							"teamB": {
								"type": "number",
								"format": "double"
							},
							"teamA": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"teamB",
							"teamA"
						],
						"type": "object"
					},
					"winner": {
						"$ref": "#/components/schemas/Winner"
					},
					"matchType": {
						"$ref": "#/components/schemas/MatchTypes"
					}
				},
				"required": [
					"title",
					"event",
					"startTime",
					"endTime",
					"teams",
					"matchType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AllEvents": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/FootballEvent"
					},
					{
						"$ref": "#/components/schemas/ChessEvent"
					},
					{
						"$ref": "#/components/schemas/CricketEvent"
					},
					{
						"$ref": "#/components/schemas/SquashMenEvent"
					},
					{
						"$ref": "#/components/schemas/SquashWomenEvent"
					},
					{
						"$ref": "#/components/schemas/TennisMenEvent"
					},
					{
						"$ref": "#/components/schemas/TennisWomenEvent"
					},
					{
						"$ref": "#/components/schemas/AthleticsEvent"
					}
				]
			},
			"AllScores": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/FootballScore"
					},
					{
						"$ref": "#/components/schemas/ChessScore"
					},
					{
						"$ref": "#/components/schemas/CricketScore"
					},
					{
						"$ref": "#/components/schemas/SquashMenScore"
					},
					{
						"$ref": "#/components/schemas/SquashWomenScore"
					},
					{
						"$ref": "#/components/schemas/TennisMenScore"
					},
					{
						"$ref": "#/components/schemas/TennisWomenScore"
					},
					{
						"$ref": "#/components/schemas/AthleticsEventScore"
					}
				]
			},
			"FootballScoreUpdateRequest": {
				"properties": {
					"teamA_score": {
						"type": "number",
						"format": "double"
					},
					"teamB_score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_score",
					"teamB_score"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "inter-iit-scoreboard-server",
		"version": "1.0.0",
		"description": "This is the server for the scoreboard for the Inter-IIT-sports-meet made by Karan Gandhi",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Karan-Gandhi"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/admin/user": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a user with the given username.",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteUserRequest"
							}
						}
					}
				}
			}
		},
		"/api/admin/createUserWithUsernameAndPassword": {
			"post": {
				"operationId": "CreateUserWithUsernameAndPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"properties": {
												"role": {
													"$ref": "#/components/schemas/UserRole"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"role",
												"name"
											],
											"type": "object"
										},
										"accessToken": {
											"type": "string"
										},
										"type": {
											"$ref": "#/components/schemas/AccessTokenTypes"
										},
										"refreshToken": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"accessToken",
										"type",
										"refreshToken"
									],
									"type": "object"
								}
							}
						}
					},
					"403": {
						"description": "Username already exists"
					}
				},
				"description": "Creates a new user with the given name, username, password, and role.",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserWithUsernameAndPasswordRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/events/chess/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a chess game.",
				"tags": [
					"Events",
					"Chess"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the chess game.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the chess game.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChessScoreUpdateRequest",
								"description": "- The updated score for the chess game."
							}
						}
					}
				}
			}
		},
		"/api/events/cricket/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a cricket match.",
				"tags": [
					"Events",
					"Cricket"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CricketScoreUpdateRequest",
								"description": "- The updated score."
							}
						}
					}
				}
			}
		},
...
		"/api/events/{id}": {
			"get": {
				"operationId": "GetEventById",
				"responses": {
					"200": {
						"description": "The event with the specified ID.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AllEvents"
								}
							}
						}
					}
				},
				"description": "Retrieves an event by its ID.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"description": "The ID of the event.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/events/updateScore/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a specific event.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the event.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the event.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AllScores",
								"description": "- The updated score for the event."
							}
						}
					}
				}
			}
		},
		"/api/events/{id}/winner": {
			"post": {
				"operationId": "SetWinner",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Sets the winner of an event.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the event.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The winner object containing the team and participants.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Winner",
								"description": "- The winner object containing the team and participants."
							}
						}
					}
				}
			}
		},
		"/api/events/{id}/vote": {
			"post": {
				"operationId": "Vote",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"action": {
										"type": "string",
										"enum": [
											"add",
											"remove"
										]
									},
									"team": {
										"type": "string",
										"enum": [
											"A",
											"B"
										]
									}
								},
								"required": [
									"action",
									"team"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/events/football/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a football match.",
				"tags": [
					"Events",
					"Football"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the football match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the match.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FootballScoreUpdateRequest",
								"description": "- The updated score for the match."
							}
						}
					}
				}
			}
		},
		"/api/global/broadcast": {
			"post": {
				"operationId": "Broadcast",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Global"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"duration": {
										"type": "number",
										"format": "double"
									},
									"message": {
										"type": "string"
									}
								},
								"required": [
									"duration",
									"message"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/global/ticker": {
			"post": {
				"operationId": "UpdateTicker",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Global"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"text": {
										"type": "string"
									}
								},
								"required": [
									"text"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/global/featured-event": {
			"post": {
				"operationId": "UpdateFeaturedEvent",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Global"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"eventId": {
										"type": "string"
									}
								},
								"required": [
									"eventId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/api/admin/schedule": {
			"get": {
				"operationId": "GetNotCompletedEvents",
				"responses": {
					"200": {
						"description": "The list of not completed events.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"$ref": "#/components/schemas/Document_any.any.any_"
											},
											{
												"$ref": "#/components/schemas/FootballEvent"
											},
											{
												"$ref": "#/components/schemas/ChessEvent"
											},
											{
												"$ref": "#/components/schemas/CricketEvent"
											},
											{
												"$ref": "#/components/schemas/SquashMenEvent"
											},
											{
												"$ref": "#/components/schemas/SquashWomenEvent"
											},
											{
												"$ref": "#/components/schemas/TennisMenEvent"
											},
											{
												"$ref": "#/components/schemas/TennisWomenEvent"
											},
											{
												"$ref": "#/components/schemas/AthleticsEvent"
											},
											{
												"$ref": "#/components/schemas/Required_any_"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves the list of not completed events.",
				"tags": [
					"Admin",
					"Schedule"
				],
				"security": [],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateExistingEvents",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates existing events.",
				"tags": [
					"Admin",
					"Schedule"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateExistingEventsRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/events/squashmen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a Squash Men match.",
				"tags": [
					"Events",
					"SquashMen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the match.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SquashMenScoreUpdateRequest",
								"description": "- The updated score for the match."
							}
						}
					}
				}
			}
		},
		"/api/events/squashwomen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a Squash Women match.",
				"tags": [
					"Events",
					"SquashWomen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the match.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SquashWomenScoreUpdateRequest",
								"description": "- The updated score for the match."
							}
						}
					}
				}
			}
		},
		"/api/admin/teams": {
			"get": {
				"operationId": "GetAllTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Team"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all teams.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Adds a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/events/tennismen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a tennis men's match.",
				"tags": [
					"Events",
					"TennisMen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the match.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TennisMenScoreUpdateRequest",
								"description": "- The updated score for the match."
							}
						}
					}
				}
			}
		},
		"/api/events/tenniswomen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a tennis women's match.",
				"tags": [
					"Events",
					"TennisWomen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the match.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TennisWomenScoreUpdateRequest",
								"description": "- The updated score for the match."
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}